move crit stuff to critvalue module
