+++ src/parser/monk/brewmaster/CombatLogParser.js
@@ -12,6 +12,7 @@ import Channeling from './modules/core/Channeling';
 import MasteryValue from './modules/core/MasteryValue';
 import AgilityValue from './modules/features/AgilityValue';
 import VersatilityValue from './modules/features/VersatilityValue';
+import CritValue from './modules/features/CritValue';
 import SpellUsable from './modules/core/SpellUsable';
 // Spells
 import IronSkinBrew from './modules/spells/IronSkinBrew';
@@ -61,6 +62,7 @@ class CombatLogParser extends CoreCombatLogParser {
     agilityValue: AgilityValue,
     masteryValue: MasteryValue,
     versValue: VersatilityValue,
+    critValue: CritValue,
     mitigationCheck: MitigationCheck,
     spellUsable: SpellUsable,
